+++ scheduler/src/cook/monitor.clj
@@ -22,6 +22,7 @@
             [cook.config :as config :refer [config]]
             [cook.datomic :as datomic]
             [cook.pool :as pool]
+            [cook.prometheus-metrics :as prometheus]
             [cook.queries :as queries]
             [cook.quota :as quota]
             [cook.scheduler.share :as share]
@@ -121,6 +122,15 @@
   (let [amount-to-inc (- (long (min value Long/MAX_VALUE)) (counters/value counter))]
     (counters/inc! counter amount-to-inc)))
 
+(defn set-prometheus-gauge!
+  "Sets the value of the counter to the new value."
+  [pool-name user state type amount]
+  ; Metrics need to be pre-registered in prometheus, so only record them if the metric exists
+  ; Log a warning otherwise so that we know to add a metric if we add a new resource type.
+  (if (contains? prometheus/resource-metric-map type)
+    (prometheus/set (prometheus/resource-metric-map type) {:pool pool-name :user user :state state} amount)
+    (log/warn "Encountered unknown type for prometheus user metrics:" type)))
+
 (defn- clear-old-counters!
   "Clears counters that were present on the previous iteration
   but not in the current iteration. This avoids the situation
@@ -133,7 +143,9 @@
         users-to-clear (difference previous-users current-users)]
     (run! (fn [user]
             (run! (fn [[type _]]
-                    (set-counter! (counters/counter [state user (name type) (str "pool-" pool-name)]) 0))
+                    (do
+                      (set-counter! (counters/counter [state user (name type) (str "pool-" pool-name)]) 0)
+                      (set-prometheus-gauge! pool-name user state type 0)))
                   (get previous-stats user)))
           users-to-clear)))
 
@@ -145,18 +157,22 @@
   (run!
     (fn [[user stats]]
       (run! (fn [[type amount]]
-              (-> [state user (name type) (str "pool-" pool-name)]
+              (do
+                (-> [state user (name type) (str "pool-" pool-name)]
                   counters/counter
-                  (set-counter! amount)))
+                  (set-counter! amount))
+                (set-prometheus-gauge! pool-name user state type amount)))
             stats))
     (add-aggregated-stats stats)))
+
 
 (defn set-total-counter!
   "Given a state (e.g. starved) and a value, sets the corresponding counter."
   [state value pool-name]
-  (-> [state "users" (str "pool-" pool-name)]
-      counters/counter
-      (set-counter! value)))
+  (do (-> [state "users" (str "pool-" pool-name)]
+        counters/counter
+        (set-counter! value))
+        (prometheus/set prometheus/user-state-count {:pool pool-name :state state} value)))
 
 (defn set-stats-counters!
   "Queries the database for running and waiting jobs per user, and sets