+++ scheduler/src/cook/prometheus_metrics.clj
@@ -44,6 +44,23 @@
 (def scheduler-filter-offensive-jobs-duration :cook/scheduler-filter-offensive-jobs-duration-seconds)
 (def scheduler-handle-status-update-duaration :cook/scheduler-handle-status-update-duaration-seconds)
 (def scheduler-handle-framework-message-duration :cook/scheduler-handle-framework-message-duration-seconds)
+(def scheduler-jobs-launched :cook/scheduler-jobs-launched-total)
+(def scheduler-match-cycle-jobs-count :cook/scheduler-match-cycle-jobs-count)
+(def scheduler-match-cycle-matched-percent :cook/scheduler-match-cycle-matched-percent)
+(def scheduler-match-cycle-head-was-matched :cook/scheduler-match-cycle-head-was-matched)
+(def scheduler-match-cycle-queue-was-full :cook/scheduler-match-cycle-queue-was-full)
+(def scheduler-match-cycle-all-matched :cook/scheduler-match-cycle-all-matched)
+(def user-state-count :cook/scheduler-users-state-count)
+;; For user resource metrics, we access them by resource type at runtime, so it is
+;; easier to define them all in a map instead of separate vars.
+(def resource-metric-map
+  {:cpus :cook/scheduler-users-cpu-count
+   :mem :cook/scheduler-users-memory-mebibytes
+   :jobs :cook/scheduler-users-jobs-count
+   :gpus :cook/scheduler-users-gpu-count
+   :launch-rate-saved :cook/scheduler-users-launch-rate-saved
+   :launch-rate-per-minute :cook/scheduler-users-launch-rate-per-minute})
+
 
 (defn create-registry
   []
@@ -116,7 +133,51 @@
                            :quantiles default-summary-quantiles})
       (prometheus/summary scheduler-handle-framework-message-duration
                           {:description "Distribution of handle framework message latency"
-                           :quantiles default-summary-quantiles}))))
+                           :quantiles default-summary-quantiles})
+      (prometheus/counter scheduler-jobs-launched
+                        {:description "Total count of jobs launched per pool and compute cluster"
+                         :labels [:pool :compute-cluster]})
+      ;; Match cycle metrics ------------------------------------------------------------------------------------
+      (prometheus/gauge scheduler-match-cycle-jobs-count
+                        {:description "Aggregate match cycle job counts stats"
+                         :labels [:pool :status]})
+      (prometheus/gauge scheduler-match-cycle-matched-percent
+                        {:description "Percent of jobs matched in last match cycle"
+                         :labels [:pool]})
+      ; The follow 1/0 metrics are useful for value map visualizations in Grafana
+      (prometheus/gauge scheduler-match-cycle-head-was-matched
+                        {:description "1 if head was matched, 0 otherwise"
+                         :labels [:pool]})
+      (prometheus/gauge scheduler-match-cycle-queue-was-full
+                        {:description "1 if queue was full, 0 otherwise"
+                         :labels [:pool]})
+      (prometheus/gauge scheduler-match-cycle-all-matched
+                        {:description "1 if all jobs were matched, 0 otherwise"
+                         :labels [:pool]})
+      ;; Resource usage stats -----------------------------------------------------------------------------------
+      ;; We set these up using a map so we can access them easily by resource type when we set the metric.
+      (prometheus/gauge (resource-metric-map :mem)
+                        {:description "Current memory by state"
+                        :labels [:pool :user :state]})
+      (prometheus/gauge (resource-metric-map :cpus)
+                        {:description "Current cpu count by state"
+                        :labels [:pool :user :state]})
+      (prometheus/gauge (resource-metric-map :gpus)
+                        {:description "Current gpu count by state"
+                        :labels [:pool :user :state]})
+      (prometheus/gauge (resource-metric-map :jobs)
+                        {:description "Current jobs count by state"
+                        :labels [:pool :user :state]})
+      (prometheus/gauge (resource-metric-map :launch-rate-saved)
+                        {:description "Current launch-rate-saved count by state"
+                        :labels [:pool :user :state]})
+      (prometheus/gauge (resource-metric-map :launch-rate-per-minute)
+                        {:description "Current launch-rate-per-minute count by state"
+                        :labels [:pool :user :state]})
+      ;; Metrics for user resource allocation counts
+      (prometheus/gauge user-state-count
+                        {:description "Current user count by state"
+                         :labels [:pool :state]}))))
 
 ;; A global registry for all metrics reported by Cook.
 ;; All metrics must be registered before they can be recorded.
@@ -128,6 +189,22 @@
   {:arglists '([name labels & body])}
   [name labels & body]
   `(prometheus/with-duration (registry ~name ~labels) ~@body))
+
+(defmacro set
+  "Sets the value of the given metric."
+  {:arglists '([name amount] [name labels amount])}
+  ([name amount]
+   `(prometheus/set registry ~name ~amount))
+  ([name labels amount]
+   `(prometheus/set registry ~name ~labels ~amount)))
+
+(defmacro inc
+  "Sets the value of the given metric."
+  {:arglists '([name] [name labels])}
+  ([name]
+   `(prometheus/inc registry ~name))
+  ([name labels]
+   `(prometheus/inc registry ~name ~labels)))
 
 (defn export []
   "Returns the current values of all registered metrics in plain text format."