+++ scheduler/src/cook/scheduler/scheduler.clj
@@ -38,7 +38,7 @@
             [cook.plugins.definitions :as plugins]
             [cook.plugins.launch :as launch-plugin]
             [cook.pool :as pool]
-            [cook.prometheus-metrics :as prometheus]
+            [cook.prometheus-metrics :as prom]
             [cook.queries :as queries]
             [cook.quota :as quota]
             [cook.rate-limit :as ratelimit]
@@ -216,8 +216,8 @@
   "Takes a status update from mesos."
   [conn pool-name->fenzo-state status]
   (log/info "Instance status is:" status)
-    (prometheus/with-duration
-      prometheus/scheduler-handle-status-update-duaration {}
+    (prom/with-duration
+      prom/scheduler-handle-status-update-duaration {}
       (timers/time!
         handle-status-update-duration
         (try (let [db (db conn)
@@ -344,8 +344,8 @@
   [conn {:keys [handle-exit-code handle-progress-message]}
    {:strs [exit-code progress-message progress-percent progress-sequence task-id] :as message}]
   (log/info "Received framework message:" {:task-id task-id, :message message})
-  (prometheus/with-duration
-    prometheus/scheduler-handle-framework-message-duration {}
+  (prom/with-duration
+    prom/scheduler-handle-framework-message-duration {}
     (timers/time!
       handle-framework-message-duration
       (try
@@ -653,8 +653,8 @@
                                              :tags {:pool pool-name :component tracing-component-tag}}]
                                          (locking fenzo
                                            (unassign-all pool-name unassigner to-unassign)
-                                           (prometheus/with-duration
-                                             prometheus/scheduler-fenzo-schedule-once-duration {:pool pool-name}
+                                           (prom/with-duration
+                                             prom/scheduler-fenzo-schedule-once-duration {:pool pool-name}
                                              (timers/time!
                                                (timers/timer (metric-title "fenzo-schedule-once-duration" pool-name))
                                                (.scheduleOnce fenzo requests leases)))))
@@ -701,8 +701,8 @@
   [unfiltered-db pool-name]
   (tracing/with-span
     [s {:name "scheduler.generate-user-usage-map" :tags {:pool pool-name :component tracing-component-tag}}]
-    (prometheus/with-duration
-      prometheus/scheduler-generate-user-usage-map-duration {:pool pool-name}
+    (prom/with-duration
+      prom/scheduler-generate-user-usage-map-duration {:pool pool-name}
       (timers/time!
         generate-user-usage-map-duration
         (->> (tools/get-running-task-ents unfiltered-db)
@@ -851,6 +851,7 @@
                 token-key (quota/pool+user->token-key pool-name user)]
             (ratelimit/spend! quota/per-user-per-pool-launch-rate-limiter token-key 1)
             (ratelimit/spend! compute-cluster-launch-rate-limiter ratelimit/compute-cluster-launch-rate-limiter-key 1))
+          (prom/inc prom/scheduler-jobs-launched {:pool pool-name :compute-cluster (cc/compute-cluster-name compute-cluster)})
           (locking fenzo
             (-> fenzo
                 (.getTaskAssigner)
@@ -949,8 +950,8 @@
                             :compute-cluster compute-cluster-name
                             :pool pool-name
                             :number-tasks count-txns})
-      (prometheus/with-duration
-        prometheus/scheduler-launch-all-matched-tasks-total-duration {:pool pool-name}
+      (prom/with-duration
+        prom/scheduler-launch-all-matched-tasks-total-duration {:pool pool-name}
         (timers/time!
           (timers/timer (metric-title "launch-matched-tasks-all-duration" pool-name))
           (try
@@ -960,8 +961,8 @@
             ;; during a race. If that happens, then other jobs that should
             ;; be scheduled will not be eligible for rescheduling until
             ;; the pending-jobs atom is repopulated
-            (prometheus/with-duration
-              prometheus/scheduler-launch-all-matched-tasks-transact-duration {:pool pool-name}
+            (prom/with-duration
+              prom/scheduler-launch-all-matched-tasks-transact-duration {:pool pool-name}
               (timers/time!
                 (timers/timer (metric-title "handle-resource-offer!-transact-task-duration" pool-name))
                 (datomic/transact
@@ -981,8 +982,8 @@
             ;; Launching the matched tasks MUST happen after the above transaction in
             ;; order to allow a transaction failure (due to failed preconditions)
             ;; to block the launch
-            (prometheus/with-duration
-              prometheus/scheduler-launch-all-matched-tasks-submit-duration {:pool pool-name}
+            (prom/with-duration
+              prom/scheduler-launch-all-matched-tasks-submit-duration {:pool pool-name}
               (timers/time!
                 (timers/timer (metric-title "handle-resource-offer!-mesos-submit-duration" pool-name))
                 (let [_ (log-structured/info "Launching matched tasks for compute cluster"
@@ -1072,8 +1073,8 @@
   [pending-jobs-for-autoscaling pool-name compute-clusters job->acceptable-compute-clusters-fn]
   (tracing/with-span
     [s {:name "scheduler.trigger-autoscaling" :tags {:component tracing-component-tag}}]
-    (prometheus/with-duration
-      prometheus/scheduler-trigger-autoscaling-duration {:pool pool-name}
+    (prom/with-duration
+      prom/scheduler-trigger-autoscaling-duration {:pool pool-name}
       (timers/time!
         (timers/timer (metric-title "trigger-autoscaling!-duration" pool-name))
         (try
@@ -1116,11 +1117,22 @@
         user->number-unmatched-considerable-jobs (merge-with
                                                    -
                                                    user->number-total-considerable-jobs
-                                                   user->number-matched-considerable-jobs)]
+                                                   user->number-matched-considerable-jobs)
+        match-percent (if (= number-considerable-jobs 0) 0 (/ number-matched-jobs number-considerable-jobs))
+        queue-was-full? (= max-considerable number-considerable-jobs)]
     (if (= number-considerable-jobs 0)
       ; keep the log slim in the 0 considerables case
-      (log-structured/info "total match cycle metric"
-                           {:inputs {:jobs-considerable 0} :pool pool-name} nil match-cycle-logger-ns)
+      (do (log-structured/info "total match cycle metric"
+                               {:inputs {:jobs-considerable 0} :pool pool-name} nil match-cycle-logger-ns)
+          ; Also reset any prometheus metrics for accuracy
+          (prom/set prom/scheduler-match-cycle-jobs-count {:pool pool-name :status "matched"} 0)
+          (prom/set prom/scheduler-match-cycle-jobs-count {:pool pool-name :status "considerable"} 0)
+          (prom/set prom/scheduler-match-cycle-jobs-count {:pool pool-name :status "unmatched"} 0)
+          (prom/set prom/scheduler-match-cycle-jobs-count {:pool pool-name :status "max-considerable"} 0)
+          (prom/set prom/scheduler-match-cycle-matched-percent {:pool pool-name} 0)
+          (prom/set prom/scheduler-match-cycle-all-matched {:pool pool-name} 0)
+          (prom/set prom/scheduler-match-cycle-head-was-matched {:pool pool-name} 0)
+          (prom/set prom/scheduler-match-cycle-queue-was-full {:pool pool-name} 0))
       ; nonzero considerables case
       (do
         ; compute the considerable, matched, and unmatched jobs for each user and emit individual metrics
@@ -1136,10 +1148,10 @@
                              {:inputs {:jobs-considerable number-considerable-jobs
                                        :offers (count offers)
                                        :max-considerable max-considerable
-                                       :queue-was-full (= max-considerable number-considerable-jobs)}
+                                       :queue-was-full queue-was-full?}
                               :matched {:jobs-considerable number-matched-jobs
                                         :offers (count offers-scheduled)
-                                        :match-percent (/ number-matched-jobs number-considerable-jobs)
+                                        :match-percent match-percent
                                         :head-was-matched head-matched?}
                               :pool pool-name
                               :unmatched {:jobs-considerable number-unmatched-jobs
@@ -1147,7 +1159,15 @@
                               :stats {:jobs-considerable (jobs->stats considerable-jobs)
                                       :offers (offers->stats offers)
                                       :head-resources head-resources}}
-                             nil match-cycle-logger-ns)))
+                             nil match-cycle-logger-ns)
+        (prom/set prom/scheduler-match-cycle-jobs-count {:pool pool-name :status "matched"} number-matched-jobs)
+        (prom/set prom/scheduler-match-cycle-jobs-count {:pool pool-name :status "considerable"} number-considerable-jobs)
+        (prom/set prom/scheduler-match-cycle-jobs-count {:pool pool-name :status "unmatched"} number-unmatched-jobs)
+        (prom/set prom/scheduler-match-cycle-jobs-count {:pool pool-name :status "max-considerable"} max-considerable)
+        (prom/set prom/scheduler-match-cycle-matched-percent {:pool pool-name} match-percent)
+        (prom/set prom/scheduler-match-cycle-all-matched {:pool pool-name} (if (= match-percent 1.0) 1 0))
+        (prom/set prom/scheduler-match-cycle-head-was-matched {:pool pool-name} (if head-matched? 1 0))
+        (prom/set prom/scheduler-match-cycle-queue-was-full {:pool pool-name} (if queue-was-full? 1 0))))
 
     (counters/inc! cycle-considerable number-considerable-jobs)
     (counters/inc! cycle-matched number-matched-jobs)
@@ -1165,21 +1185,21 @@
     ;; TODO: If there is an exception before offers are sent to fenzo (scheduleOnce) then the offers will be lost. This is fine with offer expiration, but not great.
     (tracing/with-span
       [s {:name "scheduler.handle-resource-offers" :tags {:pool pool-name :component tracing-component-tag}}]
-      (prometheus/with-duration
-        prometheus/scheduler-handle-resource-offers-total-duration {:pool pool-name}
+      (prom/with-duration
+        prom/scheduler-handle-resource-offers-total-duration {:pool pool-name}
         (timers/time!
           (timers/timer (metric-title "handle-resource-offer!-duration" pool-name))
           (try
             (let [db (db conn)
-                  considerable-jobs (prometheus/with-duration
-                                      prometheus/scheduler-handle-resource-offers-pending-to-considerable-duration {:pool pool-name}
+                  considerable-jobs (prom/with-duration
+                                      prom/scheduler-handle-resource-offers-pending-to-considerable-duration {:pool pool-name}
                                       (timers/time!
                                         (timers/timer (metric-title "handle-resource-offer!-considerable-jobs-duration" pool-name))
                                         (pending-jobs->considerable-jobs
                                           db pending-jobs user->quota user->usage num-considerable pool-name)))
                   ; matches is a vector of maps of {:hostname .. :leases .. :tasks}
-                  {:keys [matches failures]} (prometheus/with-duration
-                                               prometheus/scheduler-handle-resource-offers-match-duration {:pool pool-name}
+                  {:keys [matches failures]} (prom/with-duration
+                                               prom/scheduler-handle-resource-offers-match-duration {:pool pool-name}
                                                (timers/time!
                                                  (timers/timer (metric-title "handle-resource-offer!-match-duration" pool-name))
                                                  (match-offer-to-schedule db fenzo-state considerable-jobs offers
@@ -1189,8 +1209,8 @@
                   offers-scheduled (for [{:keys [leases]} matches
                                          lease leases]
                                      (:offer lease))
-                  matched-job-uuids (prometheus/with-duration
-                                      prometheus/scheduler-handle-resource-offers-matches-to-job-uuids-duration {:pool pool-name}
+                  matched-job-uuids (prom/with-duration
+                                      prom/scheduler-handle-resource-offers-matches-to-job-uuids-duration {:pool pool-name}
                                       (timers/time!
                                         (timers/timer (metric-title "handle-resource-offer!-match-job-uuids-duration" pool-name))
                                         (matches->job-uuids matches pool-name)))
@@ -1346,14 +1366,15 @@
                                                       (max number-unmatched-jobs)) ; Autoscale at least the pods that failed to match.
                   ;; We need to filter pending jobs based on quota so that we don't
                   ;; trigger autoscaling beyond what users have quota to actually run
-                  autoscalable-jobs (tracing/with-span [s {:name "scheduler.handle-resource-offers.generate-autoscalable-jobs"
-                                                           :tags {:pool pool-name :component tracing-component-tag}}]
-                                                       (->> pool-name
-                                                            (get @pool-name->pending-jobs-atom)
-                                                            (tools/filter-pending-jobs-for-quota pool-name (atom {}) (atom {})
-                                                                                                 user->quota user->usage (tools/global-pool-quota pool-name))
-                                                            (take max-jobs-for-autoscaling-scaled)
-                                                            (doall)))
+                  autoscalable-jobs (tracing/with-span
+                                      [s {:name "scheduler.handle-resource-offers.generate-autoscalable-jobs"
+                                          :tags {:pool pool-name :component tracing-component-tag}}]
+                                      (->> pool-name
+                                           (get @pool-name->pending-jobs-atom)
+                                           (tools/filter-pending-jobs-for-quota pool-name (atom {}) (atom {})
+                                                                                user->quota user->usage (tools/global-pool-quota pool-name))
+                                           (take max-jobs-for-autoscaling-scaled)
+                                           (doall)))
                   filtered-autoscalable-jobs (remove #(.getIfPresent caches/recent-synthetic-pod-job-uuids (:job/uuid %)) autoscalable-jobs)]
               ; When we have at least a minimum number of jobs being looked at, metric which fraction have matched.
               ; This lets us measure how well we're matching on existing resources.
@@ -1418,8 +1439,8 @@
       trigger-chan
       (fn match-jobs-event []
         (log-structured/info "Starting offer matching" {:pool pool-name})
-        (prometheus/with-duration
-          prometheus/scheduler-match-cycle-duration {:pool pool-name}
+        (prom/with-duration
+          prom/scheduler-match-cycle-duration {:pool pool-name}
           (timers/time!
             (timers/timer (metric-title "match-jobs-event" pool-name))
             (tracing/with-span
@@ -1723,8 +1744,8 @@
   (tools/chime-at-ch
     trigger-chan
     (fn cancelled-task-killer-event []
-      (prometheus/with-duration
-        prometheus/scheduler-kill-cancelled-tasks-duration {}
+      (prom/with-duration
+        prom/scheduler-kill-cancelled-tasks-duration {}
         (timers/time!
           killing-cancelled-tasks-duration
           (doseq [{:keys [db/id instance/task-id] :as task} (killable-cancelled-tasks (d/db conn))]
@@ -1796,8 +1817,8 @@
   (let [tasks (into (vec running-task-ents) pending-task-ents)
         task-comparator (tools/same-user-task-comparator tasks)
         pending-task-ents-set (into #{} pending-task-ents)
-        jobs (prometheus/with-duration
-               prometheus/scheduler-sort-jobs-hierarchy-duration {:pool pool-name}
+        jobs (prom/with-duration
+               prom/scheduler-sort-jobs-hierarchy-duration {:pool pool-name}
                (timers/time!
                  sort-jobs-duration
                  (->> tasks
@@ -1932,8 +1953,8 @@
   ;; TODO these limits should come from the largest observed host from Fenzo
   ;; .getResourceStatus on TaskScheduler will give a map of hosts to resources; we can compute the max over those
   [{max-memory-gb :memory-gb max-cpus :cpus} offensive-jobs-ch jobs]
-  (prometheus/with-duration
-    prometheus/scheduler-filter-offensive-jobs-duration {}
+  (prom/with-duration
+    prom/scheduler-filter-offensive-jobs-duration {}
     (timers/time!
       filter-offensive-jobs-duration
       (let [max-memory-mb (* 1024.0 max-memory-gb)
@@ -1984,8 +2005,8 @@
 
    It ranks the jobs by dru first and then apply several filters if provided."
   [unfiltered-db offensive-job-filter]
-  (prometheus/with-duration
-    prometheus/scheduler-rank-cycle-duration {}
+  (prom/with-duration
+    prom/scheduler-rank-cycle-duration {}
     (timers/time!
       rank-jobs-duration
       (try